+++ src/Databases/index.ts
@@ -10,6 +10,7 @@ import {
   DatabasePlan,
   DatabaseMetrics,
   DatabaseInstanceStatus,
+  DatabasePitrRestoreResponse,
   LogsArchivesResult,
   DatabaseType,
   DatabaseTypeVersion,
@@ -18,6 +19,8 @@ import {
   FirewallRuleCreateParams,
   FirewallRuleUpdateParams,
   DbOperation,
+  DatabaseUser,
+  DatabaseUserCreateParams,
 } from "../models/regional/databases";
 import { unpackData } from "../utils";
 
@@ -308,6 +311,23 @@ export default class Databases {
   }
 
   /**
+   * Restore a database to a point in time via the dbaas API
+   * @param addonId ID of the database addon
+   * @param restoreTime Point-in-time restore target
+   * @return Promise that when resolved returns the restore operation payload.
+   */
+  apiPitrRestore(
+    addonId: string,
+    restoreTime: string,
+  ): Promise<DatabasePitrRestoreResponse> {
+    return unpackData(
+      this._client.dbaasApiClient().post(`/databases/${addonId}/pitr/restore`, {
+        restore_time: restoreTime,
+      }),
+    );
+  }
+
+  /**
    * Run a database action via the dbaas API
    * @param addonId ID of the database addon
    * @param actionName Name of the action to perform
@@ -411,6 +431,68 @@ export default class Databases {
    */
   backups(databaseId: string): Backups {
     return new Backups(this._client, databaseId);
+  }
+
+  /**
+   * List all users of a database
+   * @param addonId ID of the database addon
+   * @return Promise that when resolved returns the list of database users.
+   */
+  apiDatabaseUsers(addonId: string): Promise<DatabaseUser[]> {
+    return unpackData(
+      this._client.dbaasApiClient().get(`/databases/${addonId}/users`),
+      "database-users",
+    );
+  }
+
+  /**
+   * Create a new database user
+   * @param addonId ID of the database addon
+   * @param params User creation parameters
+   * @return Promise that when resolved returns the created user.
+   */
+  apiDatabaseUserCreate(
+    addonId: string,
+    params: DatabaseUserCreateParams,
+  ): Promise<DatabaseUser> {
+    return unpackData(
+      this._client
+        .dbaasApiClient()
+        .post(`/databases/${addonId}/users`, { database_user: params }),
+      "database_user",
+    );
+  }
+
+  /**
+   * Delete a database user
+   * @param addonId ID of the database addon
+   * @param username Name of the user to delete
+   * @return Promise that resolves when the user is deleted.
+   */
+  apiDatabaseUserDelete(addonId: string, username: string): Promise<void> {
+    return unpackData(
+      this._client
+        .dbaasApiClient()
+        .delete(`/databases/${addonId}/users/${username}`),
+    );
+  }
+
+  /**
+   * Reset the password of a database user
+   * @param addonId ID of the database addon
+   * @param username Name of the user
+   * @return Promise that when resolved returns the user with new password.
+   */
+  apiDatabaseUserResetPassword(
+    addonId: string,
+    username: string,
+  ): Promise<DatabaseUser> {
+    return unpackData(
+      this._client
+        .dbaasApiClient()
+        .post(`/databases/${addonId}/users/${username}/reset_password`),
+      "database_user",
+    );
   }
 
   /**